Monitor download files update, organizing steps by page to avoid feature coupled steps, cleanup
