+++ src/test/java/gov/healthit/chpl/aqa/pageObjects/SearchPage.java
@@ -11,16 +11,6 @@ public class SearchPage {
     private WebDriver driver;
     private static WebElement element = null;
 
-    public static WebElement certIdLink1(final WebDriver driver) {
-        element = driver.findElement(By.xpath("//*[@id=\"toggle-cms-8169\"]"));
-        return element;
-    }
-
-    public static WebElement certIdLink2(final WebDriver driver) {
-        element = driver.findElement(By.xpath("//*[@id=\"toggle-cms-8969\"]"));
-        return element;
-    }
-
     /**
      * Returns the first search details button.
      *
@@ -31,7 +21,11 @@ public class SearchPage {
         element = driver.findElement(By.xpath("//*[@id=\'resultsSection\']/div[2]/div/table/tbody/tr/td[8]/div/div[1]/a"));
         return element;
     }
-
+    /**
+     * Returns Get EHR Certification ID button.
+     * @param driver webdriver
+     * @return button element
+     */
     public static WebElement getEHRCertId(final WebDriver driver) {
         element = driver.findElement(By.id("get-ehr-cert-id"));
         return element;
@@ -39,7 +33,6 @@ public class SearchPage {
 
     /**
      * Returns the pending mask "black screen".
-     *
      * @param driver WebDriver
      * @return the pending mask
      */
@@ -47,16 +40,18 @@ public class SearchPage {
         element = driver.findElement(By.id("pendingMask"));
         return element;
     }
-
+    /**
+     * Returns text no results found when search does not return results .
+     * @param driver WebDriver
+     * @return text element
+     */
     public static WebElement noResultsFound(final WebDriver driver) {
         element = driver.findElement(By.xpath("//*[@id=\"resultsSection\"]/div[2]/div/table/thead/tr/td/div/div/p"));
         return element;
     }
 
-
     /**
      * Returns the result count element.
-     *
      * @param driver webdriver
      * @return the the result count element
      */
@@ -64,7 +59,11 @@ public class SearchPage {
         element = driver.findElement(By.xpath("//*[@id=\'resultsSection\']/div[2]/div/table/thead/tr/td/div/div/div/div"));
         return element;
     }
-
+    /**
+     * Returns results section that shows search results.
+     * @param driver webdriver
+     * @return results section element
+     */
     public static WebElement resultsSection(final WebDriver driver) {
         element = driver.findElement(By.id("resultsSection"));
         return element;
@@ -72,7 +71,6 @@ public class SearchPage {
 
     /**
      * Returns the first search result's CHPL ID.
-     *
      * @param driver webdriver
      * @return the first search result's CHPL ID
      */
@@ -82,7 +80,6 @@ public class SearchPage {
     }
     /**
      * Returns main search field element.
-     *
      * @param driver webdriver
      * @return the search field element
      */
