+++ src/test/resources/Features/OCD-1954-disclosureurl.feature
@@ -1,10 +1,10 @@
-Feature: Regression test to test Mandatory Disclosures URL for listings and 
+Feature: OCD-1954 - Regression test to test Mandatory Disclosures URL for listings and 
 		 Acceptance test to test updated mandatory disclosure urls for cerner listings
 
 @Regression
 Scenario Outline: Verify Mandatory Disclosures URL exists and shows correct URL for that listing
-	Given I am on listing details page with database ID "<DB_ID>"
-	Then the Mandatory Disclosures URL field should show Disclosures URL "<URL>"
+	Given I am on listing details page of listing with database ID "<DB_ID>"
+	Then the Mandatory Disclosures URL should be: "<URL>"
 	Examples:
 		|DB_ID|URL|
 		|8624|https://www.cerner.com/cehrt-disclosure-information|
