+++ src/test/resources/Features/OCD-2062-CMSIDsLookup.feature
@@ -1,6 +1,6 @@
 @Regression
 Feature: OCD-2062 - Verify that newly generated & incorrectly removed CMS IDs still work
-  Background: Some CMS IDs were lost during an outage. They need to be re-inserted, and able to be validated against. In addition, if a new CMS ID is generated with the same CHPL Listings, the CMS ID should be the new version.
+         Some CMS IDs were lost during an outage. They need to be re-inserted, and able to be validated against. In addition, if a new CMS ID is generated with the same CHPL Listings, the CMS ID should be the new version.
 
   Scenario Outline: As an end-user, I want to be able to lookup my listings with the old CMS IDs
     Given I am on the CMS ID Reverse Lookup page
